backend/requirements.txt
uvicorn[standard]
fastapi[standard]
pydantic-settings
sqlalchemy
psycopg2
redis
openai
httpx
boto3